logs: Make command more intuitive
Since the only "logs" we get are for the cluster (rather than IdP,
ingress, etc), the user should be able to just call 'moactl logs'
directly on the cluster and that should be sufficient for the command to
understand what logs the user is requesting
