+++ docs/moactl.md
@@ -25,7 +25,7 @@ Command line tool for MOA.
 * [moactl list](moactl_list.md)	 - List all resources of a specific type
 * [moactl login](moactl_login.md)	 - Log in to your Red Hat account
 * [moactl logout](moactl_logout.md)	 - Log out
-* [moactl logs](moactl_logs.md)	 - Show logs of a specific resource
+* [moactl logs](moactl_logs.md)	 - Show cluster installation logs
 * [moactl verify](moactl_verify.md)	 - Verify resources are configured correctly for cluster install
 * [moactl version](moactl_version.md)	 - Prints the version of the tool
 * [moactl whoami](moactl_whoami.md)	 - Displays user account information
