+++ docs/moactl_logs.md
@@ -1,15 +1,32 @@
 ## moactl logs
 
-Show logs of a specific resource
+Show cluster installation logs
 
 ### Synopsis
 
-Show logs of a specific resource
+Show cluster installation logs
+
+```
+moactl logs [ID|NAME] [flags]
+```
+
+### Examples
+
+```
+  # Show last 100 log lines for a cluster named "mycluster"
+  moactl logs mycluster --tail=100
+
+  # Show logs for a cluster using the --cluster flag
+  moactl logs --cluster=mycluster
+```
 
 ### Options
 
 ```
-  -h, --help   help for logs
+  -c, --cluster string   Name or ID of the cluster to get logs for.
+  -h, --help             help for logs
+      --tail int         Number of lines to get from the end of the log. (default 2000)
+  -w, --watch            After getting the logs, watch for changes.
 ```
 
 ### Options inherited from parent commands
@@ -22,5 +39,4 @@ Show logs of a specific resource
 ### SEE ALSO
 
 * [moactl](moactl.md)	 - 
-* [moactl logs cluster](moactl_logs_cluster.md)	 - Show details of a cluster
